// filters-playground/src/interfaces.ts
export interface IOlympicData {
    athlete: string;
    age: number;
    country: string;
    year: number;
    date: string;
    sport: string;
    gold: number;
    silver: number;
    bronze: number;
    total: number;
}

export type ComplexDataType = {
    complexField: {
        value: string,
        filterValue: string,
    },
};